+++ src/main/java/org/example/express_backend/dto/CreatePackageDTO.java
@@ -6,8 +6,8 @@ import java.io.Serializable;
 
 @Data
 public class CreatePackageDTO implements Serializable {
-    private Integer shipmentId; // 运单id
-    private Integer receiverId; // 收件人id（如果在数据库中存在）
+    private Long shipmentId; // 运单id
+    private Long receiverId; // 收件人id（如果在数据库中存在）
     private String receiverName;    // 收件人姓名
     private String receiverAddress; // 收件人地址
     private String receiverPhone;   // 收件人电话
